+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.3.0 - Revival (2021/02/25)
 
+* Added Entity types
+  * Default entity selected via dropdown
+  * Per-entry entity marked in weight brackets (`ItemName{@Item}`, `ItemName{2@Item}`)
 * Added 0.8.0 support
   * Rolltable creation has changed slightly. New code should be a little more robust
 * Added BugReporter support